Add refresh paths
Now refreshes the paths in the tables so when moving to program main folder as a unit (moving the Tlamim-Back-Suggestion folder) the paths in the tables will be updated.
Notice that if we wish to rewrite the saved tables with the original paths we should export each one of the tables again (You don't have to do that).
The tables that were affected are formsTable.xlsx, generalQuestionsTable.xlsx and privateQuestionsTable.xlsx